2318: cleanup
